+++ core/physicalHealth/MetabolicEquivalentOfTask.ts
@@ -5,9 +5,12 @@ CALCULATE THE METABOLIC EQUIVALENT OF A TASK
 import CreateComponentDataUtilities from "@/core/tools/CoreLibraryDataBuilder";
 import { CoreLibraryResponse } from "@/core/types/CoreLibraryResponse";
 
-export const { getSource, getLastUpdate } = CreateComponentDataUtilities(
+export const { getSources, getLastUpdate } = CreateComponentDataUtilities(
     "01/07/2024",
-    "https://en.wikipedia.org/wiki/Metabolic_equivalent_of_task > https://www.ahajournals.org/doi/10.1161/CIRCULATIONAHA.107.185649 and https://www.inchcalculator.com/calories-burned-weight-lifting-calculator/",
+    [
+        "https://en.wikipedia.org/wiki/Metabolic_equivalent_of_task > https://www.ahajournals.org/doi/10.1161/CIRCULATIONAHA.107.185649", // TODO - further update the getSources() type to reflect related URLs
+        "https://www.inchcalculator.com/calories-burned-weight-lifting-calculator/",
+    ],
 );
 
 /**