+++ core/physicalHealth/basalMetabolicRate.ts
@@ -4,11 +4,15 @@ CALCULATE BASAL METABOLIC RATE
 
 import { CoreLibraryResponse } from "@/core/types/CoreLibraryResponse";
 import CreateComponentDataUtilities from "@/core/tools/CoreLibraryDataBuilder";
-import { CoreLibraryType_Activeness } from "../types/MiscTypes";
+import { CoreLibraryType_Activeness } from "@/core/types/MiscTypes";
 
-export const { getSource, getLastUpdate } = CreateComponentDataUtilities(
-    "26/06/2024",
-    "https://en.wikipedia.org/wiki/Harris%E2%80%93Benedict_equation and https://ajcn.nutrition.org/article/S0002-9165(23)16698-6/abstract and https://web.archive.org/web/20081014132915/http://gottasport.com/weight-loss/71/harris-benedict-formula-for-women-and-men.html",
+export const { getSources, getLastUpdate } = CreateComponentDataUtilities(
+    "04/12/2024",
+    [
+        "https://en.wikipedia.org/wiki/Harris%E2%80%93Benedict_equation",
+        "https://ajcn.nutrition.org/article/S0002-9165(23)16698-6/abstract",
+        "https://web.archive.org/web/20081014132915/http://gottasport.com/weight-loss/71/harris-benedict-formula-for-women-and-men.html",
+    ],
 );
 
 /**
@@ -26,7 +30,7 @@ export default function calculateBasalMetabolicRate(
     gender: "male" | "female",
     weight: number,
     height: number,
-    _activeness: CoreLibraryType_Activeness, // TODO - as far as I know there's an index from 1.2 to 1.whatever you've got to multiply your result by to get the real BMR. that index is based on how active you are. we are currently returning the raw BMR without taking this into account, so we aren't really being precise.
+    activeness: CoreLibraryType_Activeness,
 ): CoreLibraryResponse {
     const firstStep: number = 10 * weight;
     const secondStep: number = 6.25 * height;
@@ -40,10 +44,27 @@ export default function calculateBasalMetabolicRate(
         // female
         sixthStep = fifthStep - 161;
     }
-    const bmr: number = sixthStep;
+    let activenessIndex: number;
+    switch (activeness) {
+        case "poor":
+            activenessIndex = 1.2;
+            break;
+        case "light":
+            activenessIndex = 1.375;
+            break;
+        case "moderate":
+            activenessIndex = 1.55;
+            break;
+        case "intense":
+            activenessIndex = 1.725;
+            break;
+        case "super":
+            activenessIndex = 1.9;
+            break;
+    }
+    const bmr: number = sixthStep * activenessIndex;
 
-    const context: string | undefined =
-        "The basal metabolic rate would be " + bmr + ".";
+    const context: string = `The estimate of daily burnt calories is of ${bmr}`;
 
     const response: CoreLibraryResponse = {
         result: bmr,